create storyRoutes
