+++ server/.env.example
@@ -1,8 +1,11 @@
 ENVIRONMENT='dev'
 PORT=8080
 
-DB_HOST=localhost
-DB_USER=labber
-DB_PASSWORD=123
-DB_DATABASE=zodiac_animals
-DB_PORT=5432
+DB_HOST=
+DB_USER=
+DB_PASSWORD=
+DB_DATABASE=
+DB_PORT=
+ORIGIN=
+
+OPENAI_API_KEY=[INSERT YOUR API KEY HERE]
