+++ server/routes/animalsRoutes.js
@@ -24,20 +24,20 @@ router.get('/ai/:id', async (req, res) => {
   }
 })
 
-router.post('/create', (req, res) => {
-  const animal_id = Number(req.body.id)
-  // console.log("animal_id:", animal_id)
-  // make API call here
+// router.post('/create', (req, res) => {
+//   const animal_id = Number(req.body.id)
+//   // console.log("animal_id:", animal_id)
+//   // make API call here
   
-  .then(data => {
-    const newStory = animalsQueries.createStories(animal_id, data)
-    .then(() => {
-      res.status(200).json(newStory)
-    })
-  })
-  .catch(err => {
-    res.status(500).send("Server can't create stories", err);
-  })
-})
+//   .then(data => {
+//     const newStory = animalsQueries.createStories(animal_id, data)
+//     .then(() => {
+//       res.status(200).json(newStory)
+//     })
+//   })
+//   .catch(err => {
+//     res.status(500).send("Server can't create stories", err);
+//   })
+// })
 
 module.exports = router;
